requirements.txt
numpy
scipy
wfdb
Pillow
